refactor: 클래스명 변경
- StringCalculator -> Calculator
