+++ src/test/java/calculator/README.md
@@ -2,7 +2,6 @@
 
 ## TODO
 
-
 - split() 잘랏을때 크기 검증
 - null, 빈 값 검증
 - 사칙연산 기호 검증
